// src/enums/Permission.ts
export enum PERMISSION_TYPE {
    NORMAL = "NORMAL",
    NORMAL_SETTINGS = "NORMAL_SETTINGS",
    COMPANY_SETTINGS = "COMPANY_SETTINGS",
    SUPER_ADMIN_SETTINGS = "SUPER_ADMIN_SETTINGS",
}

export enum PERMISSION_GROUP {
    DASHBOARD = "DASHBOARD",
    PROJECTS = "PROJECTS",
    TASKS = "TASKS",
    CUSTOMERS = "CUSTOMERS",
    EXTRA = "EXTRA",
    LABELS = "LABELS",
    TASK_STATUS = "TASK_STATUS",
    SPRINTS = "SPRINTS",

    // HOTEL COMPANY SETTINGS

    BOOKING = "BOOKING",
    ROOMS = "ROOMS",
    ROOM_TYPES = "ROOM_TYPES",

    // SUPER ADMIN ONLY
    ROLES = "ROLES",
    PERMISSIONS = "PERMISSIONS",
    FEATURES = "FEATURES",
    EMPLOYEES = "EMPLOYEES",
    COMPANY = "COMPANY",
    FEATURES_AND_TEAMS = "FEATURES_AND_TEAMS",
    INTERNAL_COMPANY = "INTERNAL_COMPANY",
}

export enum PERMISSION_ENUM {
    // PROJECT
    CREATE_PROJECT = "CREATE_PROJECT",
    READ_PROJECT = "READ_PROJECT",
    UPDATE_PROJECT = "UPDATE_PROJECT",
    DELETE_PROJECT = "DELETE_PROJECT",

    // TASK
    CREATE_TASK = "CREATE_TASK",
    READ_TASK = "READ_TASK",
    UPDATE_TASK = "UPDATE_TASK",
    DELETE_TASK = "DELETE_TASK",

    // CUSTOMERS
    CREATE_CUSTOMER = "CREATE_CUSTOMER",
    READ_CUSTOMER = "READ_CUSTOMER",
    UPDATE_CUSTOMER = "UPDATE_CUSTOMER",
    DELETE_CUSTOMER = "DELETE_CUSTOMER",

    // EXTRA
    // AUDIT LOGS
    READ_AUDIT_LOGS = "READ_AUDIT_LOGS",
    // DASHBOARD
    READ_DASHBOARD = "READ_DASHBOARD",

    // LABELS
    CREATE_LABEL = "CREATE_LABEL",
    READ_LABEL = "READ_LABEL",
    UPDATE_LABEL = "UPDATE_LABEL",
    DELETE_LABEL = "DELETE_LABEL",

    // TASK_STATUS
    CREATE_TASK_STATUS = "CREATE_TASK_STATUS",
    READ_TASK_STATUS = "READ_TASK_STATUS",
    UPDATE_TASK_STATUS = "UPDATE_TASK_STATUS",
    DELETE_TASK_STATUS = "DELETE_TASK_STATUS",

    // SPRINTS
    CREATE_SPRINT = "CREATE_SPRINT",
    READ_SPRINT = "READ_SPRINT",
    UPDATE_SPRINT = "UPDATE_SPRINT",
    DELETE_SPRINT = "DELETE_SPRINT",
}

export enum SUPER_ADMIN_PERMISSION_ENUM {
    // ! SUPER ADMIN ONLY
    // EMPLOYEE
    CREATE_EMPLOYEE = "CREATE_EMPLOYEE",
    READ_EMPLOYEE = "READ_EMPLOYEE",
    READ_EMPLOYEE_DETAILS = "READ_EMPLOYEE_DETAILS",
    UPDATE_EMPLOYEE = "UPDATE_EMPLOYEE",
    DELETE_EMPLOYEE = "DELETE_EMPLOYEE",
    // ROLES
    CREATE_ROLE = "CREATE_ROLE",
    READ_ROLE = "READ_ROLE",
    UPDATE_ROLE = "UPDATE_ROLE",
    DELETE_ROLE = "DELETE_ROLE",

    // PERMISSIONS
    READ_PERMISSION = "READ_PERMISSION",
    UPDATE_PERMISSION = "UPDATE_PERMISSION",
    ASSIGN_PERMISSION = "ASSIGN_PERMISSION",

    // FEATURES_AND_TEAMS

    READ_FEATURES_AND_TEAMS = "READ_FEATURES_AND_TEAMS",
    UPDATE_FEATURES_AND_TEAMS = "UPDATE_FEATURES_AND_TEAMS",
    DELETE_FEATURES_AND_TEAMS = "DELETE_FEATURES_AND_TEAMS",
    ADD_MEMBER_TO_FEATURES_AND_TEAMS = "ADD_MEMBER_TO_FEATURES_AND_TEAMS",
    REMOVE_MEMBER_FROM_FEATURES_AND_TEAMS = "REMOVE_MEMBER_FROM_FEATURES_AND_TEAMS",

    // INTERNAL_COMPANY
    // CREATE_INTERNAL_COMPANY = "CREATE_INTERNAL_COMPANY",
    READ_INTERNAL_COMPANY = "READ_INTERNAL_COMPANY",
    UPDATE_INTERNAL_COMPANY = "UPDATE_INTERNAL_COMPANY",
    DELETE_INTERNAL_COMPANY = "DELETE_INTERNAL_COMPANY",
}

export enum COMPANY_PERMISSION_ENUM {
    // ! COMPANY ADMIN AND BELOW
    // HOTEL COMPANY SETTINGS
    // BOOKING
    CREATE_BOOKING = "CREATE_BOOKING",
    READ_BOOKING = "READ_BOOKING",
    READ_BOOKING_DETAILS = "READ_BOOKING_DETAILS",
    UPDATE_BOOKING = "UPDATE_BOOKING",
    DELETE_BOOKING = "DELETE_BOOKING",

    // ROOMS
    CREATE_ROOM = "CREATE_ROOM",
    READ_ROOM = "READ_ROOM",
    UPDATE_ROOM = "UPDATE_ROOM",
    DELETE_ROOM = "DELETE_ROOM",

    // ROOM_TYPES
    CREATE_ROOM_TYPE = "CREATE_ROOM_TYPE",
    READ_ROOM_TYPE = "READ_ROOM_TYPE",
    UPDATE_ROOM_TYPE = "UPDATE_ROOM_TYPE",
    DELETE_ROOM_TYPE = "DELETE_ROOM_TYPE",
}